Split wasm interface into its own file
