+++ src/index.ts
@@ -23,8 +23,8 @@ import { LambertGrid } from "./grids/LambertGrid";
 import { RadarSweepGrid } from "./grids/RadarSweepGrid";
 import { GeostationaryImage } from "./grids/Geostationary";
 import { UnstructuredGrid } from "./grids/UnstructuredGrid";
-
-import { initMSModule, FieldContourOpts } from './ContourCreator';
+import { initMSModule } from "./WasmInterface";
+import { FieldContourOpts } from './ContourCreator';
 
 /** All built-in colormaps */
 const colormaps = {
